Create helper file to holder clock logic and adjust useEffect hook to call helper method
